Refactor: Rename PostCard inner data classes
